==> 01 - Data Retrieval.Rmd <==
---
title: "Data Retrieval for Covid-19 Modelling"
author: "Murray McKechnie"
date: "5/14/2020"
output: html_document
---

This script retrieves the following data from various sources:

  1)  Age distribution by country
  2)  Number of coronavirus tests, cases, and fatalities by country
  3)  Obesity rates by country and gender
  4)  Overweight rates by country and gender
  5)  Government response data
  6)  GDP (PPP) for each country
  7)  Relative Deaths per 100 000 by age group, which is used to get an age-adjusted risk factor factor
  8)  Population density and spatial inequality data (calculated by Differential Capital)
  
Only (2) and (5) need to be updated regularly; the rest of the data should not have to be updated. So the code to update all the other data points has been commented out to speed up the process. However, all commented out code can be run just fine if you wish and achieves what it says it does.

(7) should also be updated if estimates of the relative risk of different age-groups change, however any new data for (7) needs to be entered in manually as this script is not able to automatically locate that data.

```{r}
rm(list = ls())
library(tidyverse)
library(httr)
```

# 1a - Age Distributions by Country

```{r}
# Population Pyramid has a separate csv for each location; url for each depends on the location ID.
# So this cycles through what appears to be the possible range of location IDs (1:1000) and loads any data it finds.
# This pics up 231 locations, mostly countries but also other regions.

# Takes ~ 15 minutes and throws a lot of red text, but that does not mean it is broken
# Cote d'Ivoire is number 384, Curacao is 531, but both have unrecognised characters in name and can't be read in so just leaving them out for now
# Monaco is number 492, contains no data and causes an error so excluded.
# I don't think there are any locations with ID > 1000, so that's where I stop.

# age_data <- list()
# 
# get_disposition_filename <- function(x) { # Function to get the file name of the file downloaded by a given url
#   sub(".*filename=", "", headers(x)$`content-disposition`)
# }
# 
# for(i in c(1:383, 385:491, 493:531, 533:637, 639:1000)) {
#   curr_data <- tryCatch(read.csv(url(paste0("https://www.populationpyramid.net/api/pp/", i, "/2019/?csv=true"))),
#            error = function(e) {NA})
#   
#   if(is.na(curr_data)) {
#     next
#   }
#   
#   hh <- HEAD(paste0("https://www.populationpyramid.net/api/pp/", i, "/2019/?csv=true")) # The name of the csv tells us what location the data is for, as well as year
#   
#   LOCATION <- gsub("[^[:alpha:]]", "", get_disposition_filename(hh)) # Only keeps location. Year always 2019
#   
#   age_data[[substr(LOCATION, 1, nchar(LOCATION)-3)]] <- curr_data # Remove "csv" from location, store data to list
# }

# If you'd rather just read it in, run this
# age_data <- readRDS(file = "01_Workspace/age_data.Rds")
```

# 1b - Get median age for each country - https://www.cia.gov/library/publications/resources/the-world-factbook/fields/343rank.html

```{r}
# median_age_raw <- read.delim(url("https://www.cia.gov/library/publications/resources/the-world-factbook/fields/rawdata_343.txt"))
# 
# median_age_raw <-  read.table(url("https://www.cia.gov/library/publications/resources/the-world-factbook/fields/rawdata_343.txt"),
#                                col.names = 1:10,
#                                fill = TRUE,
#                                skip = 1,
#                                stringsAsFactors = F)
# 
# # Unfortunately, the above data doesn't read in well, so we need to do some cleaning (mainly merging columns together). Just run this:
# apply(median_age_raw, 2, function(x) trimws(x))
# 
# median_age_raw[median_age_raw == ""] <- NA
# 
# median_age_raw$X2[is.na(median_age_raw$X5) & !is.na(median_age_raw$X4)] <- unite(median_age_raw[is.na(median_age_raw$X5) & !is.na(median_age_raw$X4), 2:3], "Country", sep = "", na.rm = T)[,1]
# median_age_raw$X3[is.na(median_age_raw$X5) & !is.na(median_age_raw$X4)] <- median_age_raw$X4[is.na(median_age_raw$X5) & !is.na(median_age_raw$X4)]
# 
# median_age_raw$X2[is.na(median_age_raw$X6) & !is.na(median_age_raw$X5)][1:10] <- unite(median_age_raw[is.na(median_age_raw$X6) & !is.na(median_age_raw$X5), 2:4], "Country", sep = "", na.rm = T)[1:10,1]
# median_age_raw$X3[is.na(median_age_raw$X6) & !is.na(median_age_raw$X5)][1:10] <- (median_age_raw$X5[is.na(median_age_raw$X6) & !is.na(median_age_raw$X5)])[1:10]
# 
# median_age_raw$X2[is.na(median_age_raw$X7) & !is.na(median_age_raw$X6)] <- unite(median_age_raw[is.na(median_age_raw$X7) & !is.na(median_age_raw$X6), 2:5], "Country", sep = "", na.rm = T)[,1]
# median_age_raw$X3[is.na(median_age_raw$X7) & !is.na(median_age_raw$X6)] <- median_age_raw$X6[is.na(median_age_raw$X7) & !is.na(median_age_raw$X6)]
# 
# median_age_raw$X2[is.na(median_age_raw$X8) & !is.na(median_age_raw$X7)] <- unite(median_age_raw[is.na(median_age_raw$X8) & !is.na(median_age_raw$X7), 2:6], "Country", sep = "", na.rm = T)[,1]
# median_age_raw$X3[is.na(median_age_raw$X8) & !is.na(median_age_raw$X7)] <- median_age_raw$X7[is.na(median_age_raw$X8) & !is.na(median_age_raw$X7)]
# 
# median_age_raw$X2[is.na(median_age_raw$X10) & !is.na(median_age_raw$X9)] <- unite(median_age_raw[is.na(median_age_raw$X10) & !is.na(median_age_raw$X9), 2:8], "Country", sep = "", na.rm = T)[,1]
# median_age_raw$X3[is.na(median_age_raw$X10) & !is.na(median_age_raw$X9)] <- median_age_raw$X9[is.na(median_age_raw$X10) & !is.na(median_age_raw$X9)]
# 
# median_age_data <- median_age_raw %>%
#   mutate(X3 = as.numeric(X3)) %>%
#   rename(Country = X2,
#          Median_Age = X3) %>%
#   select(Country, Median_Age)
```

# 2 - Load data for number of infections, fatalities, and tests for each country. Including per million people (our response variable).

This needs to be updated regularly.

```{r}
tests_cases_deaths_data <- read.csv(url("https://covid.ourworldindata.org/data/owid-covid-data.csv"))

tests_cases_deaths_data <- 
  tests_cases_deaths_data %>%
    arrange(location, date)
```

# 3 - Load obesity data for adults from WHO

```{r}
# obesity_data_raw <- read.csv(url("https://apps.who.int/gho/athena/data/GHO/NCD_BMI_30A?filter=AGEGROUP:*;COUNTRY:*;SEX:*&x-sideaxis=COUNTRY&x-topaxis=YEAR;GHO;AGEGROUP;SEX&profile=crosstable&format=csv"), stringsAsFactors = F)[,1:4]
# 
# # Clean it up a bit
# info_about_obestity_data <- obesity_data_raw[1:2,2]
# 
# obesity_data <- obesity_data_raw[-c(1:3),]
# 
# names(obesity_data) <- gsub(" ", "", obesity_data_raw[3,])
# 
# # Only keep the point estimates. Some NAs due to missing data
# obesity_data$Bothsexes <- as.numeric(substr(obesity_data$Bothsexes, 1, str_locate(obesity_data$Bothsexes, " ")[,1] - 1))
# obesity_data$Male <- as.numeric(substr(obesity_data$Male, 1, str_locate(obesity_data$Male, " ")[,1] - 1))
# obesity_data$Female <- as.numeric(substr(obesity_data$Female, 1, str_locate(obesity_data$Female, " ")[,1] - 1))
```

# 4 - Load overweight data from WHO

```{r}
# overweight_data_raw <- read.csv(url("https://apps.who.int/gho/athena/data/GHO/NCD_BMI_25A?filter=AGEGROUP:*;COUNTRY:*;SEX:*&x-sideaxis=COUNTRY&x-topaxis=YEAR;GHO;AGEGROUP;SEX&profile=crosstable&format=csv"), stringsAsFactors = F)[,1:4]
# 
# # Clean it up a bit
# info_about_overweight_data <- overweight_data_raw[1:2,2]
# 
# overweight_data <- overweight_data_raw[-c(1:3),]
# 
# names(overweight_data) <- gsub(" ", "", overweight_data_raw[3,])
# 
# # Only keep the point estimates. Some NAs due to missing data
# overweight_data$Bothsexes <- as.numeric(substr(overweight_data$Bothsexes, 1, str_locate(overweight_data$Bothsexes, " ")[,1] - 1))
# overweight_data$Male <- as.numeric(substr(overweight_data$Male, 1, str_locate(overweight_data$Male, " ")[,1] - 1))
# overweight_data$Female <- as.numeric(substr(overweight_data$Female, 1, str_locate(overweight_data$Female, " ")[,1] - 1))
```

# 5 - Government response data from Oxford - https://www.bsg.ox.ac.uk/research/research-projects/coronavirus-government-response-tracker

```{r}
gov_response_data <- read.csv(url("https://raw.githubusercontent.com/OxCGRT/covid-policy-tracker/master/data/OxCGRT_latest.csv"))
```

# 6 - GDP per capita (PPP) from IMF

```{r}
# gdp_url <- "https://www.imf.org/external/pubs/ft/weo/2019/02/weodata/weoreptc.aspx?pr.x=80&pr.y=11&sy=2020&ey=2024&scsm=1&ssd=1&sort=country&ds=.&br=1&c=512%2C668%2C914%2C672%2C612%2C946%2C614%2C137%2C311%2C546%2C213%2C674%2C911%2C676%2C314%2C548%2C193%2C556%2C122%2C678%2C912%2C181%2C313%2C867%2C419%2C682%2C513%2C684%2C316%2C273%2C913%2C868%2C124%2C921%2C339%2C948%2C638%2C943%2C514%2C686%2C218%2C688%2C963%2C518%2C616%2C728%2C223%2C836%2C516%2C558%2C918%2C138%2C748%2C196%2C618%2C278%2C624%2C692%2C522%2C694%2C622%2C962%2C156%2C142%2C626%2C449%2C628%2C564%2C228%2C565%2C924%2C283%2C233%2C853%2C632%2C288%2C636%2C293%2C634%2C566%2C238%2C964%2C662%2C182%2C960%2C359%2C423%2C453%2C935%2C968%2C128%2C922%2C611%2C714%2C321%2C862%2C243%2C135%2C248%2C716%2C469%2C456%2C253%2C722%2C642%2C942%2C643%2C718%2C939%2C724%2C734%2C576%2C644%2C936%2C819%2C961%2C172%2C813%2C132%2C726%2C646%2C199%2C648%2C733%2C915%2C184%2C134%2C524%2C652%2C361%2C174%2C362%2C328%2C364%2C258%2C732%2C656%2C366%2C654%2C144%2C336%2C146%2C263%2C463%2C268%2C528%2C532%2C923%2C944%2C738%2C176%2C578%2C534%2C537%2C536%2C742%2C429%2C866%2C433%2C369%2C178%2C744%2C436%2C186%2C136%2C925%2C343%2C869%2C158%2C746%2C439%2C926%2C916%2C466%2C664%2C112%2C826%2C111%2C542%2C298%2C967%2C927%2C443%2C846%2C917%2C299%2C544%2C582%2C941%2C474%2C446%2C754%2C666%2C698&s=PPPPC&grp=0&a="
# 
# gdp_per_capita_ppp_data <- read.delim(url(gdp_url), 1L, stringsAsFactors = F)
# 
# # Check when estimates started
# table(gdp_per_capita_ppp_data$Estimates.Start.After)
# 
# # See which countries have Estimates.Start.After < 2004
# gdp_per_capita_ppp_data %>%
#   filter(Estimates.Start.After < 2004)
# 
# # Maybe safest just to throw them away. And then just keep Country and 2020 columns
# gdp_per_capita_ppp_data <- 
#   gdp_per_capita_ppp_data %>%
#     filter(Estimates.Start.After >= 2004) %>%
#     select(Country, X2020) %>%
#     rename(gdp_estimate_ppp = X2020)
# 
# # Convert data to numeric
# gdp_per_capita_ppp_data$gdp_estimate_ppp <- as.numeric(gsub(",", "", gdp_per_capita_ppp_data$gdp_estimate_ppp))
```

# 7 - Calculate risk-adjusted age of countries using deaths per 100 000 by age group and population pyramids. We use Sweden, Netherlands and Italy's statistics.

```{r}
# # https://www.statista.com/statistics/1105061/coronavirus-deaths-by-region-in-italy/
# italy_deaths_by_age_group <- c(3, 0, 10, 53, 242, 972, 2935, 7738, 11162, 4287)
# 
# # Using the age pyramid for Italy, let's get deaths 100000 for each 10-year age-band
# deaths_per_hundred_thousand_by_age_group_italy <- 100000 * italy_deaths_by_age_group/diff(c(0, cumsum(rowSums(age_data$Italy[,2:3]))[c(seq(2, 18, 2), 21)]))
# 
# # Obtained from https://www.statista.com/statistics/1107913/number-of-coronavirus-deaths-in-sweden-by-age-groups/
# sweden_deaths_by_age_group <- c(1, 0, 7, 12, 30, 101, 235, 726, 1288, 774)
# 
# # Using the age pyramid for Sweden, let's get deaths 100000 for each 10-year age-band
# deaths_per_hundred_thousand_by_age_group_sweden <- 100000 * sweden_deaths_by_age_group / diff(c(0, cumsum(rowSums(age_data$Sweden[,2:3]))[c(seq(2, 18, 2), 21)]))
# 
# # Netherlands deaths by age group (percentages) - https://www.statista.com/statistics/1109459/coronavirus-death-casulaties-by-age-in-netherlands/
# netherlands_deaths_by_age_group <- c(0, 0, 0, 0, 0, 0.1, 0.1, 0.1, 0.1, 0.3, 0.7, 1.6, 2.8, 5.7, 10.6, 17.3, 21.3, 22, 12.7, 4.7) / 100
# 
# total_netherlands_deaths_on_6_may <- (tests_cases_deaths_data %>%
#   filter(location == "Netherlands",
#          date == "2020-05-06") %>%
#   select(total_deaths))[1,1]
# 
# # Using the age pyramid for Netherlands, let's get deaths 100000 for each 10-year age-band
# deaths_per_hundred_thousand_by_age_group_netherlands <- 100000 *
#                                                           cumsum(netherlands_deaths_by_age_group * total_netherlands_deaths_on_6_may /
#                                                           diff(c(0, cumsum(rowSums(age_data$Netherlands[,2:3]))[c(1:19, 21)])))[seq(2, 20, 2)] / 2
# 
# age_bands <- c("0-9", "10-19", "20-29", "30-39", "40-49", "50-59", "60-69", "70-79", "80-89", "90+")
# 
# # Graphs showing how the deaths per 100 000 compare. Using uniform age distribution
# age_band_midpoints <- seq(5, 95, 10)
# 
# cbind.data.frame(age_band_midpoints,
#                  deaths_per_hundred_thousand_by_age_group_italy,
#                  deaths_per_hundred_thousand_by_age_group_sweden,
#                  deaths_per_hundred_thousand_by_age_group_netherlands) %>%
#   rename(Italy = deaths_per_hundred_thousand_by_age_group_italy,
#          Sweden = deaths_per_hundred_thousand_by_age_group_sweden,
#          Netherlands = deaths_per_hundred_thousand_by_age_group_netherlands) %>%
#   mutate(Italy = Italy / sum(Italy),
#          Sweden = Sweden / sum(Sweden),
#          Netherlands = Netherlands / sum(Netherlands)) %>%
#   pivot_longer(cols = c("Italy", "Sweden", "Netherlands")) %>%
#   ggplot(aes(x = age_band_midpoints, y = value, color = name)) +
#   geom_point() +
#   geom_line() +
#   scale_x_continuous(breaks = age_band_midpoints, labels = age_bands) +
#   labs(title = "Likelihood of Age Group given Death",
#        subtitle = "If each age group contains same number of people",
#        x = "Age group",
#        y = "Probability")
# 
# cbind.data.frame(age_band_midpoints,
#                  deaths_per_hundred_thousand_by_age_group_italy,
#                  deaths_per_hundred_thousand_by_age_group_sweden,
#                  deaths_per_hundred_thousand_by_age_group_netherlands) %>%
#   rename(Italy = deaths_per_hundred_thousand_by_age_group_italy,
#          Sweden = deaths_per_hundred_thousand_by_age_group_sweden,
#          Netherlands = deaths_per_hundred_thousand_by_age_group_netherlands) %>%
#   mutate(Italy = Italy / sum(Italy),
#          Sweden = Sweden / sum(Sweden),
#          Netherlands = Netherlands / sum(Netherlands)) %>%
#   pivot_longer(cols = c("Italy", "Sweden", "Netherlands")) %>%
#   filter(age_band_midpoints <= 60) %>%
#   ggplot(aes(x = age_band_midpoints, y = value, color = name)) +
#   geom_point() +
#   geom_line() +
#   scale_x_continuous(breaks = age_band_midpoints, labels = age_bands) +
#   labs(title = "Likelihood of Age Group given Death",
#        subtitle = "If each age group contains same number of people",
#        x = "Age group",
#        y = "Probability")
# 
# 
# # Sweden and Netherlands show very similar results, Italy slightly different results. Let's just take the average of the 3.
# 
# relative_fatality_rates <- rowMeans(cbind.data.frame(deaths_per_hundred_thousand_by_age_group_italy,
#                                                      deaths_per_hundred_thousand_by_age_group_sweden,
#                                                      deaths_per_hundred_thousand_by_age_group_netherlands))
# 
# 
# population_fatality_rate <- c()
# 
# for(i in names(age_data)) {
#   temp_data <- as.data.frame(age_data[i])
# 
#   entry <- sum(relative_fatality_rates *
#                  diff(c(0, cumsum(rowSums(temp_data[,2:3]))[c(seq(2, 18, 2), 21)])) /
#                  sum(rowSums(temp_data[,2:3])))
# 
#   population_fatality_rate <- c(population_fatality_rate, entry)
# 
# }
# 
# # Let's use Venezuela as the comparison point as it has a nice median age of 30.0. This is
# risk_adjusted_age <- cbind.data.frame(as.character(names(age_data)),
#                                               population_fatality_rate * 30 / population_fatality_rate[names(age_data) == "Venezuela"])
# 
# names(risk_adjusted_age) <- c("Country", "age_adjusted_risk")
# 
# risk_adjusted_age$Country <- as.character(risk_adjusted_age$Country)
```

# 8 - Read in density parameters obtained by Differential Capital

The inequality field is a measure of how unequally distributed the population density is in a country. It is calculated using the population density maps from https://data.humdata.org/organization/facebook.

```{r}
# density_data <- read.csv("00_Local_data/country_density_dat.csv")
```

# Save data

```{r}
# saveRDS(age_data, file = "01_Workspace/age_data.Rds")
# saveRDS(density_data, "01_Workspace/density_data.Rds")
# saveRDS(gdp_per_capita_ppp_data, "01_Workspace/gdp_per_capita_ppp_data.Rds")
saveRDS(gov_response_data, "01_Workspace/gov_response_data.Rds")
# saveRDS(median_age_data, "01_Workspace/median_age_data.Rds")
# saveRDS(obesity_data, "01_Workspace/obesity_data.Rds")
# saveRDS(overweight_data, "01_Workspace/overweight_data.Rds")
# saveRDS(risk_adjusted_age, "01_Workspace/risk_adjusted_age.Rds")
saveRDS(tests_cases_deaths_data, "01_Workspace/tests_cases_deaths_data.Rds")
```
