==> 02 - Data Cleaning.Rmd <==
---
title: "02 - Data Cleaning and Merging for Covid-19 Modelling"
author: "Murray McKechnie"
date: "5/14/2020"
output: html_document
---

In 01, we retrieved the data and cleaned it up to some extent. But now we need to merge the different data sources into one. 

This mainly involves fixing discrepencies in country names

```{r, include=F}
knitr::opts_chunk$set(echo = TRUE)

rm(list = ls())
library(lubridate)
library(tidyverse)
```

# Load Objects from 1st script

```{r}
age_data <- readRDS("01_Workspace/age_data.Rds")
density_data <- readRDS("01_Workspace/density_data.Rds")
gdp_per_capita_ppp_data <- readRDS("01_Workspace/gdp_per_capita_ppp_data.Rds")
gov_response_data <- readRDS("01_Workspace/gov_response_data.Rds")
median_age_data <- readRDS("01_Workspace/median_age_data.Rds")
obesity_data <- readRDS("01_Workspace/obesity_data.Rds")
overweight_data <- readRDS("01_Workspace/overweight_data.Rds")
risk_adjusted_age <- readRDS("01_Workspace/risk_adjusted_age.Rds")
tests_cases_deaths_data <- readRDS("01_Workspace/tests_cases_deaths_data.Rds")
```

# Find and correct the country name discrepencies between different data sources

For now, we're just merging age data, obesity data, gdp data, tests_cases_deaths data, government response data, and density data

## Join tests_cases_deaths data to obesity data

```{r}
tests_cases_deaths_data$location <- gsub(" ", "", tests_cases_deaths_data$location)
obesity_data$Country <- gsub(" ", "", obesity_data$Country)
gdp_per_capita_ppp_data$Country <- gsub(" ", "", gdp_per_capita_ppp_data$Country)

# Make obesity_data names correspond to tests_cases_deaths_data names

# See what the differences are
# setdiff(obesity_data$Country, tests_cases_deaths_data$location)
# setdiff(tests_cases_deaths_data$location, obesity_data$Country)

# Fix them
obesity_data$Country[obesity_data$Country == "Bolivia(PlurinationalStateof)"] <- "Bolivia"
obesity_data$Country[obesity_data$Country == "Venezuela(BolivarianRepublicof)"] <- "Venezuela"
obesity_data$Country[obesity_data$Country == "UnitedKingdomofGreatBritainandNorthernIreland"] <- "UnitedKingdom"
obesity_data$Country[obesity_data$Country == "UnitedStatesofAmerica"] <- "UnitedStates"
obesity_data$Country[obesity_data$Country == "DemocraticRepublicoftheCongo"] <- "DemocraticRepublicofCongo"
obesity_data$Country[obesity_data$Country == "Côted'Ivoire"] <- "Coted'Ivoire"
obesity_data$Country[obesity_data$Country == "BruneiDarussalam"] <- "Brunei"
obesity_data$Country[obesity_data$Country == "CaboVerde"] <- "CapeVerde"
obesity_data$Country[obesity_data$Country == "Czechia"] <- "CzechRepublic"
obesity_data$Country[obesity_data$Country == "DemocraticPeople'sRepublicofKorea"] <- "SouthKorea"
obesity_data$Country[obesity_data$Country == "Eswatini"] <- "Swaziland"
obesity_data$Country[obesity_data$Country == "Iran(IslamicRepublicof)"] <- "Iran"
obesity_data$Country[obesity_data$Country == "LaoPeople'sDemocraticRepublic"] <- "Laos"
obesity_data$Country[obesity_data$Country == "RepublicofMoldova"] <- "Moldova"
obesity_data$Country[obesity_data$Country == "RepublicofNorthMacedonia"] <- "Macedonia"
obesity_data$Country[obesity_data$Country == "RussianFederation"] <- "Russia"
obesity_data$Country[obesity_data$Country == "SyrianArabRepublic"] <- "Syria"
obesity_data$Country[obesity_data$Country == "Timor-Leste"] <- "Timor"
obesity_data$Country[obesity_data$Country == "UnitedRepublicofTanzania"] <- "Tanzania"
obesity_data$Country[obesity_data$Country == "VietNam"] <- "Vietnam"
obesity_data$Country[obesity_data$Country == "RepublicofKorea"] <- "NorthKorea"

# Join the objects
simple_model_data <- full_join(tests_cases_deaths_data %>%
                                 select(location, date,
                                        total_cases, new_cases, total_cases_per_million, new_cases_per_million,
                                        total_deaths, new_deaths, total_deaths_per_million, new_deaths_per_million,
                                        total_tests, new_tests, total_tests_per_thousand, new_tests_per_thousand,
                                        population) %>%
                                 rename(Country = location, population_size = population),
                               obesity_data %>% select(Country, Bothsexes) %>% rename(Obesity_rate = Bothsexes))
```

## Add in median age data

```{r}
# Find differences
# setdiff(simple_model_data$Country, median_age_data$Country)[order(setdiff(simple_model_data$Country, median_age_data$Country))]
# setdiff(median_age_data$Country, simple_model_data$Country)[order(setdiff(median_age_data$Country, simple_model_data$Country))]

# Fix them
simple_model_data$Country[simple_model_data$Country == "FaeroeIslands"] <- "FaroeIslands"
simple_model_data$Country[simple_model_data$Country == "UnitedStatesVirginIslands"] <- "VirginIslands"
simple_model_data$Country[simple_model_data$Country == "Micronesia(FederatedStatesof)"] <- "Micronesia"

median_age_data$Country[median_age_data$Country == "Bahamas,The"] <- "Bahamas"
median_age_data$Country[median_age_data$Country == "CaboVerde"] <- "CapeVerde"
median_age_data$Country[median_age_data$Country == "Congo,DemocraticRepublicofthe"] <- "DemocraticRepublicofCongo"
median_age_data$Country[median_age_data$Country == "Congo,Republicofthe"] <- "Congo"
median_age_data$Country[median_age_data$Country == "Czechia"] <- "CzechRepublic"
median_age_data$Country[median_age_data$Country == "Eswatini"] <- "Swaziland"
median_age_data$Country[median_age_data$Country == "Gambia,The"] <- "Gambia"
median_age_data$Country[median_age_data$Country == "Korea,North"] <- "NorthKorea"
median_age_data$Country[median_age_data$Country == "Korea,South"] <- "SouthKorea"
median_age_data$Country[median_age_data$Country == "Micronesia,FederatedStatesof"] <- "Micronesia"
median_age_data$Country[median_age_data$Country == "Timor-Leste"] <- "Timor"

simple_model_data <- full_join(simple_model_data, median_age_data)
```

## Add in GDP per capita data

```{r}
# Find differences
# setdiff(simple_model_data$Country, gdp_per_capita_ppp_data$Country)[order(setdiff(simple_model_data$Country, gdp_per_capita_ppp_data$Country))]
# setdiff(gdp_per_capita_ppp_data$Country, simple_model_data$Country)[order(setdiff(gdp_per_capita_ppp_data$Country, simple_model_data$Country))]

# Fix them
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "BruneiDarussalam"] <- "Brunei"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "CaboVerde"] <- "CapeVerde"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "C\U3e34663cted'Ivoire"] <- "Coted'Ivoire"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "Eswatini"] <- "Swaziland"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "HongKongSAR"] <- "HongKong"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "IslamicRepublicofIran"] <- "Iran"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "Korea"] <- "SouthKorea"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "KyrgyzRepublic"] <- "Kyrgyzstan"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "LaoP.D.R."] <- "Laos"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "MacaoSAR"] <- "Macau"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "NorthMacedonia"] <- "Macedonia"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "RepublicofCongo"] <- "Congo"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "SlovakRepublic"] <- "Slovakia"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "St.KittsandNevis"] <- "SaintKittsandNevis"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "St.Lucia"] <- "SaintLucia"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "St.VincentandtheGrenadines"] <- "SaintVincentandtheGrenadines"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "TaiwanProvinceofChina"] <- "Taiwan"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "TheBahamas"] <- "Bahamas"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "TheGambia"] <- "Gambia"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "S\U3e33653coTom\U3e39653candPr\U3e64653cncipe"] <- "SaoTomeAndPrincipe"
gdp_per_capita_ppp_data$Country[gdp_per_capita_ppp_data$Country == "Timor-Leste"] <- "Timor"

simple_model_data <- full_join(simple_model_data, gdp_per_capita_ppp_data)
```

## Add in the age_adjusted_risk factor

```{r}
# Find differences
# setdiff(simple_model_data$Country, population_fatality_rates_relative_to_Italy$Country)[order(setdiff(simple_model_data$Country, population_fatality_rates_relative_to_Italy$Country))]
# setdiff(population_fatality_rates_relative_to_Italy$Country, simple_model_data$Country)[order(setdiff(population_fatality_rates_relative_to_Italy$Country, simple_model_data$Country))]

# Fix them
risk_adjusted_age$Country[risk_adjusted_age$Country == "BruneiDarussalam"] <- "Brunei"
risk_adjusted_age$Country[risk_adjusted_age$Country == "CaboVerde"] <- "CapeVerde"
risk_adjusted_age$Country[risk_adjusted_age$Country == "ChinaMacaoSAR"] <- "Macau"
risk_adjusted_age$Country[risk_adjusted_age$Country == "ChinaHongKongSAR"] <- "HongKong"
risk_adjusted_age$Country[risk_adjusted_age$Country == "DemPeoplesRepublicofKorea"] <- "NorthKorea"
risk_adjusted_age$Country[risk_adjusted_age$Country == "DRCongo"] <- "DemocraticRepublicofCongo"
risk_adjusted_age$Country[risk_adjusted_age$Country == "GuineaBissau"] <- "Guinea-Bissau"
risk_adjusted_age$Country[risk_adjusted_age$Country == "RepublicofKorea"] <- "SouthKorea"
risk_adjusted_age$Country[risk_adjusted_age$Country == "RepublicofMoldova"] <- "Moldova"
risk_adjusted_age$Country[risk_adjusted_age$Country == "RussianFederation"] <- "Russia"
risk_adjusted_age$Country[risk_adjusted_age$Country == "StateofPalestine"] <- "Palestine"
risk_adjusted_age$Country[risk_adjusted_age$Country == "SyrianArabRepublic"] <- "Syria"
risk_adjusted_age$Country[risk_adjusted_age$Country == "TFYRMacedonia"] <- "Macedonia"
risk_adjusted_age$Country[risk_adjusted_age$Country == "UnitedStatesofAmerica"] <- "UnitedStates"
risk_adjusted_age$Country[risk_adjusted_age$Country == "UnitedStatesVirginIslands"] <- "VirginIslands"
risk_adjusted_age$Country[risk_adjusted_age$Country == "VietNam"] <- "Vietnam"
risk_adjusted_age$Country[risk_adjusted_age$Country == "TimorLeste"] <- "Timor"

simple_model_data <- full_join(simple_model_data, risk_adjusted_age)
```

## Calculate population of each country by comparing total cases to total cases per million - not longer required as test_cases_deaths data actually has population_size!

```{r}
# simple_model_data <- simple_model_data[!is.na(simple_model_data$Country),]
# simple_model_data <- simple_model_data[order(simple_model_data$Country),]
# 
# population_size <- c()
# i = "SouthAfrica"
# for(i in unique(simple_model_data$Country)) {
#   temp_data <- simple_model_data %>% filter(Country == i)
#   
#   if(sum(!is.na(temp_data$total_cases)) == 0 |
#      max(temp_data$total_cases) == 0 |
#      sum(!is.na(temp_data$total_cases_per_million)) == 0) {
#     population_size <- c(population_size, rep(0, length(temp_data$total_cases)))
#   } else(
#     population_size <- c(population_size, rep(max(temp_data$total_cases, na.rm = T) /
#                                                 max(temp_data$total_cases_per_million, na.rm = T),
#                                               length(temp_data$total_cases)))
#   )
# }
# 
# sum(is.na(population_size))
# 
# simple_model_data <- cbind.data.frame(simple_model_data, population_size)
# 
# simple_model_data %>% group_by(Country, population_size) %>% distinct(Country, population_size)
```

## Add in density data

```{r}
names(density_data)[2] <- "Country"
names(density_data)[6] <- "Density_km"
density_data$Country <- gsub(" ", "", density_data$Country)

# Find differences
# setdiff(density_data$Country, simple_model_data$Country)
# setdiff(simple_model_data$Country, density_data$Country)

# Fix them
density_data$Country[density_data$Country == "Bahamas,The"] <- "Bahamas"
density_data$Country[density_data$Country == "FederatedStatesofMicronesia"] <- "Micronesia"
density_data$Country[density_data$Country == "Gambia,The"] <- "Gambia"
density_data$Country[density_data$Country == "IvoryCoast"] <- "Coted'Ivoire"
density_data$Country[density_data$Country == "PacificIslands(Palau)"] <- "Palau"
density_data$Country[density_data$Country == "St.KittsandNevis"] <- "SaintKittsandNevis"
density_data$Country[density_data$Country == "St.Lucia"] <- "SaintLucia"
density_data$Country[density_data$Country == "St.VincentandtheGrenadines"] <- "SaintVincentandtheGrenadines"
density_data$Country[density_data$Country == "Tanzania,UnitedRepublicof"] <- "Tanzania"
density_data$Country[density_data$Country == "WesternSamoa"] <- "Samoa"

density_data$Inequality_spatial_transformed <- 1 / (1 - density_data$inequality)

simple_model_data <-
  simple_model_data %>%
    left_join(density_data %>%
                select(Country, area, Density_km, Inequality_spatial_transformed))
```

## Add in government response data

```{r}
gov_response_data$CountryName <- gsub(" ", "", gov_response_data$CountryName)

names(gov_response_data)[c(1, 3)] <- c("Country", "date")

# Find Differences
# setdiff(unique(gov_response_data$Country), unique(simple_model_data$Country))
# setdiff(unique(simple_model_data$Country), unique(gov_response_data$Country))

# Fix them
gov_response_data$Country[gov_response_data$Country == "Macao"] <- "Macau"
gov_response_data$Country[gov_response_data$Country == "KyrgyzRepublic"] <- "Kyrgyzstan"
gov_response_data$Country[gov_response_data$Country == "SlovakRepublic"] <- "Slovakia"
gov_response_data$Country[gov_response_data$Country == "Eswatini"] <- "Swaziland"

gov_response_data$date <- ymd(gov_response_data$date)

simple_model_data$date <- as.Date(simple_model_data$date)

simple_model_data <- 
  simple_model_data %>%
    left_join(gov_response_data %>%
                select(-CountryCode, -ConfirmedCases, -ConfirmedDeaths, -StringencyIndex) %>%
                rename(StringencyIndex = StringencyIndexForDisplay))

```

```{r}
# Tidy up simple_model_data a bit
simple_model_data <- simple_model_data[!is.na(simple_model_data$total_cases),]

simple_model_data <- simple_model_data[order(simple_model_data$Country),]

simple_model_data$date <- as.Date(simple_model_data$date)
```

# Save important objects

```{r}
saveRDS(simple_model_data, "02_Workspace/simple_model_data.Rds")
saveRDS(tests_cases_deaths_data, "02_Workspace/tests_cases_deaths_data.Rds")
```
