==> 03 - Determining days since initial outbreak.Rmd <==
---
title: "03 - Alligning countries at different stages of "
author: "Murray"
date: "5/9/2020"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

rm(list = ls())
library(caTools)
library(GGally)
library(tidyverse)
```

# Load data

```{r}
simple_model_data <- readRDS("02_Workspace/simple_model_data.Rds")
tests_cases_deaths_data <- readRDS("02_Workspace/tests_cases_deaths_data.Rds")
```

# For each country, get: date of first case, date (and gov response) when total_deaths_per_million crosses 0.1, 10 deaths data

```{r}
sum(is.na(simple_model_data$total_cases))

simple_model_data[is.na(simple_model_data$total_cases),]

simple_model_data <- simple_model_data[!is.na(simple_model_data$total_cases),]

simple_model_data <- simple_model_data[order(simple_model_data$Country),]

simple_model_data$date <- as.Date(simple_model_data$date)

# Start date

start_date_by_country <- 
  as.vector(t((simple_model_data %>%
    filter(total_cases > 0) %>%
    group_by(Country) %>%
    filter(max(total_deaths_per_million) >= 0.1) %>%
    slice(1) %>%
    ungroup() %>%
    select(date))[,1]))

# ten_deaths_date

ten_deaths_date_by_country <- 
  as.vector(t((simple_model_data %>%
    filter(total_deaths_per_million >= 0.1) %>%
    group_by(Country) %>%
    mutate(date = case_when(max(total_deaths, na.rm = T) < 10 ~ as.Date("9999-12-31"),
                            TRUE ~ date)) %>%
    ungroup() %>%
    filter(date == "9999-12-31" | total_deaths >= 10) %>%
    group_by(Country) %>%
    slice(1) %>%
    ungroup() %>%
    select(date))[,1]))

ten_deaths_date_by_country[ten_deaths_date_by_country == "9999-12-31"] <- NA

# CD/M > 0.1

significant_deaths_start_date_by_country <- 
  as.vector(t((simple_model_data %>%
    filter(total_deaths_per_million >= 0.1) %>%
    group_by(Country) %>%
    slice(1) %>%
    ungroup() %>%
    select(date))[,1]))

# gov_response at time where CD/M goes past 0.1

gov_response_snapshot_by_country <- 
  as.vector(t((simple_model_data %>%
    filter(total_deaths_per_million >= 0.1) %>%
    group_by(Country) %>%
    slice(1) %>%
    ungroup() %>%
    select(StringencyIndex))[,1]))

group_sizes <- as.vector(t((simple_model_data %>%
    filter(total_cases > 0) %>%
    group_by(Country) %>%
    filter(max(total_deaths_per_million) >= 0.1) %>%
    summarise(n = n()))[,2]))

simple_model_data <- simple_model_data %>%
    filter(total_cases > 0) %>%
    group_by(Country) %>%
    filter(max(total_deaths_per_million) >= 0.1) %>%
    ungroup() %>%
    mutate(days_since_outbreak = date - as.Date(rep(start_date_by_country, group_sizes)),
           days_since_deaths_crossed_0.1_per_million = date - as.Date(rep(significant_deaths_start_date_by_country, group_sizes)),
           days_since_10_deaths = date - as.Date(rep(ten_deaths_date_by_country, group_sizes)),
           gov_response_at_0.1_deaths_per_million = rep(gov_response_snapshot_by_country, group_sizes))

```

# Add rolling 5-day averages to data and add a flag for if peak has been reached based on the moving average.

```{r}
deaths_5_day_averages <- c()
infections_5_day_averages <- c()
peak_deaths_already_occurred_indicator <- c()
peak_cases_already_occurred_indicator <- c()

for(i in unique(tests_cases_deaths_data$location)) {
  temp_data <- tests_cases_deaths_data %>% filter(location == i)
  
  temp_deaths_ra <- runmean(temp_data$new_deaths_per_million, k = 5, align = "right")
  temp_cases_ra <- runmean(temp_data$new_cases_per_million, k = 5, align = "right")
  
  deaths_5_day_averages <- c(deaths_5_day_averages, temp_deaths_ra)
  infections_5_day_averages <- c(infections_5_day_averages, temp_cases_ra)
  
  if(length(na.omit(temp_deaths_ra)) == 0) {
    peak_deaths_already_occurred_indicator <- c(peak_deaths_already_occurred_indicator, rep(0, length(temp_deaths_ra)))
  } else {
    peak_deaths_already_occurred_indicator <- c(peak_deaths_already_occurred_indicator,
                                                case_when(which.max(temp_deaths_ra) < length(temp_deaths_ra) - 3 ~ rep(1, length(temp_deaths_ra)),
                                                          TRUE ~ rep(0, length(temp_deaths_ra))))
    
  }
  
  if(length(na.omit(temp_cases_ra)) == 0) {
    peak_cases_already_occurred_indicator <- c(peak_cases_already_occurred_indicator, rep(0, length(temp_cases_ra)))
  } else {
  peak_cases_already_occurred_indicator <- c(peak_cases_already_occurred_indicator,
                                             case_when(which.max(temp_cases_ra) < length(temp_cases_ra) - 3 ~ rep(1, length(temp_cases_ra)),
                                                       TRUE ~ rep(0, length(temp_cases_ra))))
    
  }
}

tests_cases_deaths_data_2 <- tests_cases_deaths_data %>%
  rename(Country = location) %>%
  cbind.data.frame(deaths_5_day_averages, infections_5_day_averages, peak_deaths_already_occurred_indicator, peak_cases_already_occurred_indicator)

simple_model_data <- left_join(simple_model_data,
                               tests_cases_deaths_data_2 %>%
                                 select(Country,
                                        date,
                                        deaths_5_day_averages,
                                        infections_5_day_averages,
                                        peak_deaths_already_occurred_indicator,
                                        peak_cases_already_occurred_indicator,
                                        new_deaths_per_million,
                                        new_cases_per_million) %>%
                                 mutate(date = as.Date(date)))
```

# Save Objects for part 4

```{r}
saveRDS(simple_model_data, "03_Workspace/simple_model_data.Rds")
```




# Look at the curves for countries that might make it into the model, see if their peaks we've found make sense

```{r}
# Let's select 10 random countries to compare

random_countries <- sample(unique(simple_model_data$Country), 1)

simple_model_data_2 <- simple_model_data %>%
                         filter(population_size >= 3 &
                                peak_deaths_already_occurred_indicator == 1) %>%
                         group_by(Country) %>%
                         filter(max(total_deaths) >= 50) %>%
                         ungroup()

death_curves <- list()
death_curves_by_threshold <- list()

# canada_data <- simple_model_data_2 %>% filter(Country == "Canada")
# 
# ggplot(data = canada_data,
#        aes(x = days_since_outbreak, y = runmean(new_deaths_per_million, 5))) +
#   geom_line()

for(i in unique(simple_model_data_2$Country)) {
  death_curves[[i]] <- ggplot(data = simple_model_data_2 %>% filter(Country == i),
                              aes(x = days_since_outbreak, y = new_deaths_per_million, color = paste0(Country, " Deaths"))) +
    geom_line() +
    geom_point() +
    # geom_line(data = simple_model_data_2 %>% filter(Country == i), 
    #           aes(x = days_since_outbreak, y = new_cases_per_million, color = paste0(Country, " Cases"))) +
    geom_line(data = simple_model_data_2 %>% filter(Country == i),
              aes(x = days_since_outbreak, y = runmean(new_deaths_per_million, 5, align = "right"), color = "5-day rolling average")) +
    # geom_line(data = simple_model_data_2 %>% filter(Country == i),
    #           aes(x = days_since_outbreak, y = runmean(new_cases_per_million, 5, align = "right"), color = "5-day rolling average")) +
    # geom_vline(xintercept = as.integer((simple_model_data_2 %>%
    #              group_by(Country) %>%
    #              filter(Country == i & runmean(new_cases_per_million, 5, align = "right") == max(runmean(new_cases_per_million, 5, align = "right"), na.rm = T)) %>%
    #              ungroup() %>%
    #              select(days_since_outbreak))[1,1])[1], linetype = "dashed") +
    geom_vline(xintercept = as.integer((simple_model_data_2 %>%
                 group_by(Country) %>%
                 filter(Country == i & runmean(new_deaths_per_million, 5, align = "right") == max(runmean(new_deaths_per_million, 5, align = "right"), na.rm = T)) %>%
                 ungroup() %>%
                 select(days_since_outbreak))[1,1])[1], linetype = "dotted") +
    theme(legend.title = element_blank())
  
    death_curves_by_threshold[[i]] <- ggplot(data = simple_model_data_2 %>% filter(Country == i),
                              aes(x = days_since_deaths_crossed_0.1_per_million, y = new_deaths_per_million, color = Country)) +
    geom_line() +
    geom_point() +
    geom_line(data = simple_model_data_2 %>% filter(Country == i),
              aes(x = days_since_deaths_crossed_0.1_per_million, y = runmean(new_deaths_per_million, 5, align = "right"), color = "5-day rolling average"))
}

# Comparing these two graphs, it seems days_since_deaths_crossed_0.1_per_million brings peaks closer together, so is probably preferable
simple_model_data_2 %>%
  group_by(Country) %>%
  filter(deaths_5_day_averages == max(deaths_5_day_averages)) %>%
  ggplot(aes(x = days_since_outbreak)) +
    geom_density()

simple_model_data_2 %>%
  group_by(Country) %>%
  filter(deaths_5_day_averages == max(deaths_5_day_averages)) %>%
  ggplot(aes(x = days_since_deaths_crossed_0.1_per_million)) +
    geom_density()

# simple_model_data_2 %>%
#   group_by(Country) %>%
#   mutate(event = case_when(deaths_5_day_averages == max(deaths_5_day_averages) ~ "Deaths_peak",
#                            infections_5_day_averages == max(infections_5_day_averages) ~ "Cases_peak",
#                            TRUE ~ "Nothing interesting")) %>%
#   filter((deaths_5_day_averages == max(deaths_5_day_averages) |
#          infections_5_day_averages == max(infections_5_day_averages)) &
#          peak_deaths_already_occurred_indicator == 1 &
#          peak_cases_already_occurred_indicator == 1) %>%
#   select(Country, date, event) %>%
#   group_by(Country, event) %>%
#   slice(1) %>%
#   ungroup() %>%
#   pivot_wider(id_cols = Country, values_from = date, names_from = event) %>%
#   ggplot(aes(x = Deaths_peak - Cases_peak)) +
#     geom_density(bw = 3)
```

```{r}
death_curves[1:21]
```

```{r}
death_curves[22:42]
```

```{r}
death_curves_by_threshold[1:21]
```

```{r}
death_curves_by_threshold[22:42]
```
